refactor(tui): split reducer tests by interaction
